MFU - Icemans further improvements
add UL-C device side read card with authentication
add MF_UL-Annotations
add ntag, and more ul descriptions in hf mfu info

+++ client/cmdhfmfu.h
@@ -19,7 +19,7 @@ void rol (uint8_t *data, const size_t len);
 
 uint8_t requestAuthentication( uint8_t *nonce);
 int try3DesAuthentication( uint8_t *key);
-uint8_t GetHF14AMfU_Type(void);
+uint16_t GetHF14AMfU_Type(void);
 
 int usage_hf_mfu_dump(void);
